+++ tests/test_table.py
@@ -1662,12 +1662,31 @@ class TestMerge(AgateTestCase):
 
         self.assertIsNot(table_c, table_a)
         self.assertIsNot(table_c, table_b)
-        self.assertColumnNames(table_c, self.column_names)
-        self.assertColumnTypes(table_c, [Number, Number, Text])
-        self.assertRows(table_c, self.rows + self.rows)
+        self.assertColumnNames(table_c, self.column_names + column_names)
+        self.assertColumnTypes(table_c, [Number, Number, Text, Number, Number, Text])
+        self.assertSequenceEqual(table_c.rows[0], [1, 4, 'a', None, None, None])
+        self.assertSequenceEqual(table_c.rows[3], [None, None, None, 1, 4, 'a'])
 
         for row in table_c.rows:
-            self.assertSequenceEqual(row.keys(), self.column_names)
+            self.assertSequenceEqual(row.keys(), self.column_names + column_names)
+
+    def test_merge_mixed_names(self):
+        table_a = Table(self.rows, self.column_names, self.column_types)
+
+        column_names = ['two', 'one', 'four']
+
+        table_b = Table(self.rows, column_names, self.column_types)
+        table_c = Table.merge([table_a, table_b])
+
+        self.assertIsNot(table_c, table_a)
+        self.assertIsNot(table_c, table_b)
+        self.assertColumnNames(table_c, ['one', 'two', 'three', 'four'])
+        self.assertColumnTypes(table_c, [Number, Number, Text, Text])
+        self.assertSequenceEqual(table_c.rows[0], [1, 4, 'a', None])
+        self.assertSequenceEqual(table_c.rows[3], [4, 1, None, 'a'])
+
+        for row in table_c.rows:
+            self.assertSequenceEqual(row.keys(), ['one', 'two', 'three', 'four'])
 
     def test_merge_different_types(self):
         table_a = Table(self.rows, self.column_names, self.column_types)
@@ -1676,8 +1695,8 @@ class TestMerge(AgateTestCase):
 
         table_b = Table(self.rows, self.column_names, column_types)
 
-        with self.assertRaises(ValueError):
-            table_c = Table.merge([table_a, table_b])  # noqa
+        with self.assertRaises(DataTypeError):
+            table_c = Table.merge([table_a, table_b])
 
     def test_merge_with_row_names(self):
         table_a = Table(self.rows, self.column_names, self.column_types, row_names='three')