+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 1.2.2
 -----
 
+* :meth:`.Table.merge` can now merge tables with different columns or columns in a different order. (#465)
+* :meth:`.MappedSequence.get` will no longer raise ``KeyError`` if a default is not provided. (#467)
 * :class:`.Number` can now test/cast the ``long`` type on Python 2.
 
 1.2.1 - February 5, 2016